perf(terrain): halve the maxima with half floats
The max pyramid is a third of the clipmap's memory and it is a chain of
maxima -- a quantity that only ever needs to be an upper bound, not an
accurate one. `R16Float` costs 298 MiB less at the shipped 4096 window,
taking the whole clipmap from 1493 MiB to 1195.

What matters is which way it rounds. A cell rounded *down* is no longer
a ceiling: a ray comparing against one skips a cell it should have
descended into, and the ground it passes through is a ridge with a hole
in it -- the failure the closed-square reduction exists to prevent, and
the reason the reduction is per-quad in the first place. So
`ceiling_half` converts to nearest and then, if that landed below the
value, steps one bit towards positive infinity. Half floats carry eleven
bits of mantissa, so rounding up costs a metre or two at the height of a
mountain and far less near sea level; all it buys the march is the
occasional descent into a cell it could have skipped.

The nodata sentinel survives, which is the other thing worth checking.
Rounding -32767 towards positive infinity gives -32736, still far below
`NODATA_BELOW`, so a cell with nothing but holes under it still reads as
one rather than as ground thirty kilometres down.

The per-level scalar the far field takes its "above everything and
climbing" early out from is now read back out of the half floats rather
than kept at full precision. Either would be a valid bound; taking it
from what was actually stored means the scalar and the texture cannot
say different things about the same window.

`half` is added as a direct dependency. It was already in the lock file
as a transitive one, so it costs no new build.

Measured against `assets/terrain` at a 1024 window, where the runs are
quick enough to repeat:

    texture   120 MiB -> 96 MiB
    draw      89.8 ms -> 90.2 ms at 6000 m, 86.0 -> 75.7 low
    frame     two pixels of 518 400 differ, by one of 255

At the shipped 4096 window the same change is 1493 MiB -> 1195 MiB. The
GPU test that checks every cell against the raster now asserts a range
rather than equality -- at or above the ground it covers, and no further
above it than one half-float step -- so a conversion that rounded the
wrong way, or that rounded further than it had to, both fail.

+++ src/terrain/clipmap.rs
@@ -114,11 +114,12 @@ impl Default for ClipmapConfig {
             // moving the frame by 0.11 of 255; eight sheds only a twelfth, and
             // two starts to show where the ring blend and the march disagree.
             near_rings: 4.0,
-            // What a 4096 window costs on the raster this flies: seven levels
-            // of heights, colours and maxima. Sized to admit that rather than
-            // to any round number, because the window is the one knob worth
-            // spending memory on and the sizes it can take are powers of two --
-            // a budget between two of them buys nothing.
+            // Room for a 4096 window on the raster this flies, which is seven
+            // levels of heights, colours and maxima and comes to 1195 MiB.
+            // Sized to admit that rather than to any round number, because the
+            // window is the one knob worth spending memory on and the sizes it
+            // can take are powers of two -- a budget between two of them buys
+            // nothing.
             memory_budget: 1600 << 20,
             // 1080p at sixty degrees, replaced wherever a real viewport is
             // known.
@@ -197,14 +198,14 @@ impl ClipmapConfig {
     /// Bytes of texture a clipmap of this shape occupies.
     ///
     /// Heights and colours are four bytes a texel at full window size; the max
-    /// pyramid is a real mip chain, so it costs about a third again over its
-    /// own base.
+    /// pyramid is two bytes and a real mip chain, so it costs about a third
+    /// again over its own base.
     pub fn texture_bytes(&self, levels: u32) -> usize {
         let window = self.window_texels as usize;
         let mut per_level = window * window * (size_of::<f32>() + 4);
         for mip in 0..=self.max_mip() {
             let side = (window >> mip).max(1);
-            per_level += side * side * size_of::<f32>();
+            per_level += side * side * size_of::<u16>();
         }
         per_level * levels as usize
     }
